configuraciones básicas terminadas para la autenticación

+++ config/jwt.php
@@ -120,8 +120,10 @@ return [
     |
     */
 
-    'refresh_ttl' => env('JWT_REFRESH_TTL', 20160),
+    //'refresh_ttl' => env('JWT_REFRESH_TTL', 20160),
+    'refresh_ttl' => (int) env('JWT_REFRESH_TTL', 20160),
 
+    
     /*
     |--------------------------------------------------------------------------
     | JWT hashing algorithm
